+++ internal/provider/api_key_resource.go
@@ -6,7 +6,7 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/cockroachdb/cockroach-cloud-sdk-go/v8/pkg/client"
+	"github.com/cockroachdb/cockroach-cloud-sdk-go/v9/pkg/client"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"