+++ internal/provider/cluster_tracking.go
@@ -26,7 +26,7 @@ import (
 	"strings"
 	"sync"
 
-	"github.com/cockroachdb/cockroach-cloud-sdk-go/v8/pkg/client"
+	"github.com/cockroachdb/cockroach-cloud-sdk-go/v9/pkg/client"
 )
 
 const clusterTrackingEnvVar = "CLEANUP_TRACKING_FILE"